src/requirements.txt
boto3
pytz
google-api-python-client
google-auth
google-auth-httplib2
